requirements.txt
customtkinter
beautifulsoup4
urllib3
requests
pillow
psutil
markdown2
tkinterweb
cloudscraper
selenium
yt-dlp
gallery-dl